backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
websockets==12.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-jose[cryptography]==3.3.0
python-multipart==0.0.6
aiofiles==23.2.1
aiohttp==3.9.1
redis==5.0.1
psutil==5.9.6
PyJWT==2.8.0
